requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets>=10.0,<12.0
python-dotenv>=0.19.0
openai==1.3.5
Pillow==10.1.0
flask>=2.0.0
cozepy>=0.13.0
httpx>=0.27.2
dashscope>=0.1.0